Make the Push abstraction more stateful
The "Push" API can now be initialized before the push is found; it maintains
internal state (`Ref`) that it uses to track the push.

This means we can "initialize" the push API _without_ actually finding the push
when a patcher is initialized: this is important, because finding the Push
requires the use of the Max for Live LiveAPI interface and that is not possible
until the device is fully initialized.